pyproject.toml
[project]
name = "mealie-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.0",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "textual>=6.2.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.9.0",
    "vulture>=2.14",
]

[tool.ruff]
# Set the maximum line length to 120 characters
line-length = 120

# Assume Python 3.13
target-version = "py313"

# Exclude common directories
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
# Enable recommended rules plus additional quality checks
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "Q",      # flake8-quotes
    "RET",    # flake8-return
    "PTH",    # flake8-use-pathlib
    "ASYNC",  # flake8-async
    "C90",    # mccabe complexity
]

# Disable specific rules that may be too strict
ignore = [
    "E501",   # Line too long (handled by formatter)
    "SIM102", # Use a single if statement (sometimes multiple ifs are clearer)
    "SIM108", # Use ternary operator (sometimes if/else is clearer)
]

# Allow autofix for all enabled rules
fixable = ["ALL"]
unfixable = []

# Set maximum complexity (cyclomatic complexity)
mccabe.max-complexity = 10

[tool.ruff.lint.per-file-ignores]
# Allow unused imports in __init__.py files
"__init__.py" = ["F401"]
# Allow assertions and print in test files
"tests/**/*.py" = ["S101", "T201"]

[tool.ruff.lint.isort]
# Use Google-style import organization
force-single-line = false
force-sort-within-sections = false
lines-after-imports = 2
known-first-party = ["mealie_parser"]

[tool.ruff.lint.pydocstyle]
# Use NumPy-style docstrings
convention = "numpy"

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"

# Use spaces for indentation
indent-style = "space"

# Respect magic trailing comma
skip-magic-trailing-comma = false

# Automatically format docstrings
docstring-code-format = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
